[GITHUB-11] add integration tests
What does this PR do?
Addresses #11

This PR adds a variety of test files for various shapes to test the overall service end to end on. It also adds an IntegrationTests class that runs all the test files and confirms that the result matches what is expected.

How has this PR been tested?
It is a test.